Validate subordinate block manifests
Each block has a subordinate block manifest; this is a collection of block
hashes which occurred in the subordinate chain from which this block was
produced. Since a dominant chain cannot validate data from a subordinate chain,
he must instead ask the subordinate (if he has a trusted subordinate to ask) if
the manifest is valid. Only then can the manifest be used to build consensus on
confirmed ETXs from external chains.

This patch accomplishes this, by passing the manifest down to our subordinate
node (iff we have a trusted subordinate), so that the subordinate node can check
validity of the dom's manifest before appending the block. Essentially:
1) dom checks manifest against manifestHash in the header
2) dom gives manifest hash to sub when asking sub to append the block
3) sub calculates expected manifest and checks against the hash provided by dom

If any check fails, the dom or sub will throw an error, preventing _both_ from
accepting the block.