ReceiveMinedHeader() needs to reconstruct pending body with sub manifest
The pending header which a miner begins working on, originates from a zone node.
Since each zone will have a different manifest, according to its own chain's
progression, the region and prime nodes can not have known what the SubManifest
would be at the time they generated their pending headers. For that reason, when
a miner submits a mined header of region or prime difficulty, the region and
prime nodes need to query their subordinates to get the correct subordinate
block manifest for that header, and reconstruct their local block bodies with
this manifest.