Add ETXs to inbound set
As ETXs become referenced through coincident block manifests, our dom will share
them with us. We need to add them to an inbound ETX set, so that they are
available to be mined into a block. This patch also implements expiry, so that
the inbound ETX set won't grow without bound.