Added GetHorizon method in Headerchain and using it to sync
GetHorizon method gives the number of the N-20 prime block. Using this to
sync and also during the block fetch will help with the sync of fray
instead of sticking with the current header which can change by quite a
bit on a dom reorg.

Changed the futureHeaderCacheLimit to 1800.

This is the link to the track the issue.
https://github.com/dominant-strategies/go-quai/issues/543.