Updated the phCache update logic
Main objective of this commit is to decouple the construction of
pendingHeader, Adding pendingHeader to the phCache and setting the
pendingHeader hash(i.e, the best knowledge pendingHeader key).

In this created 3 functions,
1) computePendingHeader - computes the pendingHeader
2) writePh - write the pendingHeader to the phCache
3) pickPhCacheHead - picks the pendingHeader head hash

Another change that was made is that for the construction of the
pendingHeader now the domPendingHeader is passed, so the synchronous
construction happens during the append process and there is no
pendingHeader return that is being sent as the response to the Append.

Also to prevent mining on a stale pendingHeader, when the pick of the
pendingHeader happends, the update is only sent to the miner if the new
pendingHeader is better than the old value in the phCache for the given
termini.

Lastly, changed the relayPh to relay current pending header,
removed genesis case in updatePhCacheFromDom,
renamed twist to cyclic reference