Fetching the parent if not found in block fetcher
During the handling of the block announces from our peer, if the peer
sends a block that is 50 blocks ahead and has a higher number than the
horizon (3 prime blocks back), we try appending the block. This is done
to ensure that the fray is synched and it increases the reorg tolerance
due to the increase in optionalites of the head.