bugfix: ETX set needs to be written for a block hash, even if the block fails append
Since the local pending header is computed before calling subclient Append(), we
need to construct the header under the optimistic assumption that the
subordinate client will accept the block, even though we do not know that to be
true. Since the worker needs the latest ETX set to build the next block, this
means the ETX set must be written directly to the database after processing the
block, instead of as part of the batch of writes to be processed after the block
has been accepted.