Refactor of phCache and head picking to always choose the greatest Td in zones:
This is a departure from the top-down picking originally proposed by HLCR. Only zones
have sufficient information to make decisions in which block to keep mining.  The
hierarchical constraint is now statistically enforced by setting subordinate Tds to
dominant values on coincident blocks.  Note that this is a statistical enforcement
and not a strict enforcement as previously proposed. This appears to create a much
more robust head pick that never stalls or seperates.

Coordinate updates coerce the ph pick to the result of the coordinate coincident
dom head choice.  The dom and its head choice is neccassarily the correct choice
because the the dom is always shared.

Renamed phHashKey, added deepcopies around phcache writes, combined writePh and pickPh into writePhHeaderAndPickPhHead