bugfix: Reduce EtxSet storage from O(n^2) to O(n)
In order to process a block with inbound ETXs, we need to know the
EtxSet (list of referencable ETXs) as of the parent block. This means,
we must be able to look up the EtxSet for any given block hash.

Previously, we naively stored the entire EtxSet for each block. This is
highly innefficent though, because a single ETX may sit in the EtxSet
for many blocks without being spent. Every time we store an EtxSet, we
are writing duplicates of ETXs that were already written in prior
EtxSets. Storage requirements grow O(n^2) with block height.

This patch modifies it, so that we only store a list of ETX hashes at
each block, and the ETXs themselves are only stored once.