Reduce ETX expiration age
Since we take a snapshot of the EtxSet after every block, the longer an
ETX stays in the EtxSet, the more its storage requirements compound.
Since we do not yet have pruning/gc for EtxSets, we need to reduce the
ETX expiration age for now. One day, pruning will obviate the need for
this.