Increased the pending body limit in worker from 32 to 320
Right now this number acts as our reorg tolerance number but in the future
after recovery and checkpointing feature, we will have the ability to generate
pending header at any point at the worst case. This can also be mitigated by
writing this body to disk and then garbage collecting it after.