Neutered the pre fetcher in the worker
After enabling snaps, we were seeing a lot of crashes in the pre fetcher and
was happening because of the pre fetcher started in the worker. Removing it for
now and also since worker does not work on multiple state objects anymore,
there is less need for the prefetcher. If we think that this becomes a
performance issue we can add it back in at a point where it doesnt crash the
node