bugfix: Removed the cache from the GetBlock/GetBody/GetHeader/OrCandidate call
There exists race conditions where once a header is read out of the Cache and
it gets evicted from the cache, node sees nil pointer dereference. This was
observed many times while testing on Garden environment on slices. (slices will
experience this more because the cache alloted for slices is smaller in size to
save on RAM usage).